refactor: 승자 구하는 구문 WinnerRepository로 위임
